修改实体类数字相关的数据类型，解决Hibernate自动生成外键出错的问题，以及完成了实验室信息的开发

+++ src/main/java/cn/edu/yangtzeu/service/DepartmentService.java
@@ -18,7 +18,7 @@ public class DepartmentService {
         return departmentDao.findAll();
     }
 
-    public Department findOne(short departmentId) {
+    public Department findOne(int departmentId) {
         return departmentDao.findOne(departmentId);
     }
